Job process contacts module ready.

+++ app/Http/Controllers/JobProcessController.php
@@ -7,6 +7,7 @@ use App\JobProcess;
 use App\JobProcessLog;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
+use Illuminate\Support\Facades\Auth;
 use Symfony\Component\HttpFoundation\Response as HttpCodes;
 
 class JobProcessController extends JsonController
@@ -18,6 +19,10 @@ class JobProcessController extends JsonController
 
     public function viewJobProcess(int $id): Response
     {
+        if (!JobProcess::isJobProcessFromUser($id, Auth::user()->id)) {
+            return static::buildUnauthorizedResponse();
+        }
+
         return static::buildResponse(JobProcess::query()->findOrFail($id)->first()->toPublicList());
     }
 
@@ -44,6 +49,10 @@ class JobProcessController extends JsonController
 
     public function editJobProcess(int $id, Request $request): Response
     {
+        if (!JobProcess::isJobProcessFromUser($id, Auth::user()->id)) {
+            return static::buildUnauthorizedResponse();
+        }
+
         $jobProcessValidator = JobProcess::getValidatorForEditPayload($request->json()->all());
 
         if ($jobProcessValidator->fails()) {
@@ -65,27 +74,10 @@ class JobProcessController extends JsonController
 
     public function deleteJobProcess(int $id): Response
     {
+        if (!JobProcess::isJobProcessFromUser($id, Auth::user()->id)) {
+            return static::buildUnauthorizedResponse();
+        }
+
         return static::buildResponse(JobProcess::deleteProcess($id));
-    }
-
-    public function changeJobProcessStatus(int $id, Request $request): Response
-    {
-        $jobProcessStatusValidator = JobProcessLog::getValidatorForChangeStatusPayload($request->json()->all());
-
-        if ($jobProcessStatusValidator->fails()) {
-            return static::buildResponse(
-                ['errors' => $jobProcessStatusValidator->errors()->getMessageBag()->toArray()],
-                HttpCodes::HTTP_UNPROCESSABLE_ENTITY
-            );
-        }
-
-        try {
-            return static::buildResponse(
-                JobProcessLog::changeJobProcessStatus($id, $jobProcessStatusValidator->validated()['job_process_status_id']),
-                HttpCodes::HTTP_CREATED
-            );
-        } catch (\Exception $ex) {
-            return static::buildResponse($ex->getMessage(), HttpCodes::HTTP_BAD_REQUEST);
-        }
     }
 }
